victoria-consumer/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
python-dateutil==2.8.2
